Clean BRFSS
BRFSS was edited in 2018 to remove SOGI (Sexual Orientation and Gender Identity) information. I cleaned the data the rest of the way up and applied linear regression to fill in the data that was missing as best as possible. This dataset would probably be more useful for validating the health scores of individual states instead of logging population data.